Integrate clean cache operation into HttpProxyCacheServer; remove extra code from sample
