+++ library/src/main/java/com/danikula/videocache/StorageUtils.java
@@ -7,6 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
+import java.io.IOException;
 
 import static android.os.Environment.MEDIA_MOUNTED;
 
@@ -80,4 +81,39 @@ final class StorageUtils {
         }
         return appCacheDir;
     }
+
+    public static void cleanDirectory(File file) {
+        if (!file.exists()) {
+            return;
+        }
+        File[] contentFiles = file.listFiles();
+        if (contentFiles != null) {
+            for (File contentFile : contentFiles) {
+                try {
+                    delete(contentFile);
+                } catch (IOException e) {
+                    LOG.warn(e.getMessage());
+                }
+            }
+        }
+    }
+
+    private static void delete(File file) throws IOException {
+        if (file.isFile() && file.exists()) {
+            deleteOrThrow(file);
+        } else {
+            cleanDirectory(file);
+            deleteOrThrow(file);
+        }
+    }
+
+    private static void deleteOrThrow(File file) throws IOException {
+        if (file.exists()) {
+            boolean isDeleted = file.delete();
+            if (!isDeleted) {
+                throw new IOException(String.format("File %s can't be deleted", file.getAbsolutePath()));
+            }
+        }
+    }
+
 }
