+++ sample/src/main/java/com/danikula/videocache/sample/App.java
@@ -5,6 +5,8 @@ import android.content.Context;
 
 import com.danikula.videocache.HttpProxyCacheServer;
 
+import java.io.File;
+
 /**
  * @author Alexey Danilov (danikula@gmail.com).
  */
@@ -19,7 +21,7 @@ public class App extends Application {
 
     private HttpProxyCacheServer newProxy() {
         return new HttpProxyCacheServer.Builder(this)
-                .cacheDirectory(Utils.getVideoCacheDir(this))
+                .cacheDirectory(new File(getExternalCacheDir(), "video-cache"))
                 .build();
     }
 }
