--- sample/src/main/java/com/danikula/videocache/sample/Utils.java
@@ -1,53 +1,0 @@
-package com.danikula.videocache.sample;
-
-import android.content.Context;
-
-import java.io.File;
-import java.io.IOException;
-
-/**
- * Some utils methods.
- *
- * @author Alexey Danilov (danikula@gmail.com).
- */
-public class Utils {
-
-    public static File getVideoCacheDir(Context context) {
-        return new File(context.getExternalCacheDir(), "video-cache");
-    }
-
-    public static void cleanVideoCacheDir(Context context) throws IOException {
-        File videoCacheDir = getVideoCacheDir(context);
-        cleanDirectory(videoCacheDir);
-    }
-
-    private static void cleanDirectory(File file) throws IOException {
-        if (!file.exists()) {
-            return;
-        }
-        File[] contentFiles = file.listFiles();
-        if (contentFiles != null) {
-            for (File contentFile : contentFiles) {
-                delete(contentFile);
-            }
-        }
-    }
-
-    private static void delete(File file) throws IOException {
-        if (file.isFile() && file.exists()) {
-            deleteOrThrow(file);
-        } else {
-            cleanDirectory(file);
-            deleteOrThrow(file);
-        }
-    }
-
-    private static void deleteOrThrow(File file) throws IOException {
-        if (file.exists()) {
-            boolean isDeleted = file.delete();
-            if (!isDeleted) {
-                throw new IOException(String.format("File %s can't be deleted", file.getAbsolutePath()));
-            }
-        }
-    }
-}
